requirements.txt
aiogram
groq
python-dotenv
yt-dlp
